separate visualization helpers from plan helpers
